+++ src/main/java/de/retest/web/selenium/css/PredicateFactory.java
@@ -1,0 +1,12 @@
+package de.retest.web.selenium.css;
+
+import java.util.function.Predicate;
+
+import de.retest.recheck.ui.descriptors.Element;
+
+@FunctionalInterface
+public interface PredicateFactory {
+
+	Predicate<Element> create( String attribute );
+
+}