+++ src/main/java/de/retest/web/selenium/css/Transformer.java
@@ -1,0 +1,8 @@
+package de.retest.web.selenium.css;
+
+@FunctionalInterface
+interface Transformer {
+
+	Selector transform( String selector );
+
+}